♻️ [Refactoring]: #304 [FE] ローカル今日日付を Due.todayLocal に集約し DueBadge / CalendarView から委譲
DueBadge と CalendarView に同一実装で重複していた todayLocal を
Due companion に Due.todayLocal として 1 箇所へ集約し、両コンポーネントから参照する。
fake timer によるローカル日付の境界テストを追加。外部挙動は不変。
